create book operation class
